Refactor packages - move webscrapers to package com.java.academy.
