+++ src/main/java/com/java/academy/controller/MainController.java
@@ -2,6 +2,7 @@ package com.java.academy.controller;
 
 import com.java.academy.model.Book;
 import com.java.academy.service.BookService;
+import com.java.academy.webScrappers.ravelo.RaveloBookProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -9,7 +10,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
-import webscrapper.raveloScrapper.RaveloBookProvider;
 
 import java.util.List;
