+++ src/main/java/com/java/academy/webScrappers/BookScrapper.java
@@ -1,17 +1,12 @@
-package webscrapper;
+package com.java.academy.webScrappers;
 
 public interface BookScrapper {
     String getBookAuthor();
-
     String getImageUrl();
-
     String getBookCategory();
-
     String getBookTitle();
-
     String getBookLink();
-
     Double getBookPrice();
-
     String getDiscount();
 }
+
