+++ src/main/resources/hibernate.properties
@@ -3,7 +3,7 @@ jdbc.url                = jdbc:mysql://localhost:3306/sys?UseUnicode=true&charac
 jdbc.user.name          = root
 jdbc.password           = kasiula 911
 
-hibernate.hbm2ddl.auto  = create
+hibernate.hbm2ddl.auto  = update
 hibernate.show_sql      = true
 
 hibernate.format_sql          = true
